+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## 12/01/2022 3.2.4
 
-- GenerateExamples - Support single generation rule
+- Generate Examples - Support specified generation rule
 - API Scenario
   - Support file type in formdata and body
   - Change the prefix length to 8 from 10 in generated API Scenario